web/requirements.txt
django
django-jet
djangorestframework
requests
psycopg2